Add a safe wrapper around ARec
* The `ARec` newtype constructor was exposed by `Data.Vinyl.ARec`,
  making it easy to break type safety. Add a `Data.Vinyl.ARec.Internal`
  module and make `Data.Vinyl.ARec` export the type abstractly. Add
  a role annotation to prevent `coerce` from breaking the abstraction.

* Hack around doctests failing to compile. I don't know the "right"
  way to do this, but something seems better than nothing.
